\name{isoboost-package}
\alias{isoboost-package}
\alias{isoboost}
\docType{package}
\title{
Isotonic Boosting Classification Rules
}
\description{
In this package we present new boosting classification rules based on LogitBoost when it can be assumed that higher (or lower) values of some predictors are related to higher levels of the response.
}
\details{
Package: isoboost

Type: Package

Version: 1.0.2

Date: 2026-01-27

License: GPL-2 | GPL-3


For a complete list of functions with individual help pages, use \code{library(help = "isoboost")}.
}
\author{
David Conde, Miguel A. Fernandez, Cristina Rueda, Bonifacio Salvador

Maintainer: David Conde <daconrio@gmail.com>
}

\keyword{ package }
\keyword{ multivariate }
\keyword{ classif }
\keyword{ tree }
